+++ packages/client/base/src/types/embed/v3/CrossmintEmbeddedCheckoutV3Props.ts
@@ -28,13 +28,25 @@ export type CrossmintEmbeddedCheckoutV3Props =
 
 export type EmbeddedCheckoutV3Recipient = EmbeddedCheckoutV3EmailRecipient | EmbeddedCheckoutV3WalletAddressRecipient;
 
+export type EmbeddedCheckoutV3PhysicalAddress = {
+    name: string;
+    line1: string;
+    line2?: string;
+    city: string;
+    state: string;
+    postalCode: string;
+    country: "US";
+};
+
 export type EmbeddedCheckoutV3EmailRecipient = {
     email: string;
     walletAddress?: never;
+    physicalAddress?: EmbeddedCheckoutV3PhysicalAddress;
 };
 export type EmbeddedCheckoutV3WalletAddressRecipient = {
     walletAddress: string;
     email?: never;
+    physicalAddress?: EmbeddedCheckoutV3PhysicalAddress;
 };
 
 export type EmbeddedCheckoutV3LineItem =
@@ -51,6 +63,9 @@ export type EmbeddedCheckoutV3LineItem =
           tokenLocator: string;
           callData?: never;
           executionParameters?: Record<string, any>;
+      }
+    | {
+          productLocator: string;
       };
 
 export type EmbeddedCheckoutV3Appearance = {