+++ packages/client/wallets/smart-wallet/src/blockchain/wallets/account/strategy.ts
@@ -1,5 +1,0 @@
-import type { AccountAndSigner, WalletCreationContext } from "../../../types/internal";
-
-export interface AccountCreationStrategy {
-    create(params: WalletCreationContext): Promise<AccountAndSigner>;
-}