+++ apps/auth/node-and-express-server/CHANGELOG.md
@@ -1,5 +1,17 @@
 # crossmint-auth-node
 
+## 1.1.19
+
+### Patch Changes
+
+- @crossmint/server-sdk@1.2.8
+
+## 1.1.18
+
+### Patch Changes
+
+- @crossmint/server-sdk@1.2.7
+
 ## 1.1.17
 
 ### Patch Changes