+++ apps/payments/nextjs/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @crossmint/client-sdk-nextjs-starter
 
+## 1.4.33
+
+### Patch Changes
+
+- Updated dependencies [98d8197]
+  - @crossmint/client-sdk-react-ui@1.15.0
+
+## 1.4.32
+
+### Patch Changes
+
+- @crossmint/client-sdk-base@1.4.11
+- @crossmint/client-sdk-react-ui@1.14.9
+
+## 1.4.31
+
+### Patch Changes
+
+- Updated dependencies [b1d7f0c]
+  - @crossmint/client-sdk-react-ui@1.14.8
+
+## 1.4.30
+
+### Patch Changes
+
+- c83894c: checkout: product locator types
+- Updated dependencies [c83894c]
+  - @crossmint/client-sdk-base@1.4.10
+  - @crossmint/client-sdk-react-ui@1.14.7
+
 ## 1.4.29
 
 ### Patch Changes