+++ apps/payments/vue/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @crossmint/client-sdk-vue-ui-starter
 
+## 1.1.47
+
+### Patch Changes
+
+- Updated dependencies [98d8197]
+  - @crossmint/client-sdk-smart-wallet@0.2.0
+
+## 1.1.46
+
+### Patch Changes
+
+- @crossmint/client-sdk-base@1.4.11
+- @crossmint/client-sdk-vue-ui@1.1.35
+- @crossmint/client-sdk-smart-wallet@0.1.36
+
+## 1.1.45
+
+### Patch Changes
+
+- Updated dependencies [c83894c]
+  - @crossmint/client-sdk-base@1.4.10
+  - @crossmint/client-sdk-vue-ui@1.1.34
+  - @crossmint/client-sdk-smart-wallet@0.1.35
+
 ## 1.1.44
 
 ### Patch Changes