+++ packages/client/verifiable-credentials/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @crossmint/client-sdk-verifiable-credentials
 
+## 3.4.26
+
+### Patch Changes
+
+- @crossmint/client-sdk-base@1.4.11
+
+## 3.4.25
+
+### Patch Changes
+
+- Updated dependencies [c83894c]
+  - @crossmint/client-sdk-base@1.4.10
+
 ## 3.4.24
 
 ### Patch Changes