+++ packages/server/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @crossmint/server-sdk
 
+## 1.2.8
+
+### Patch Changes
+
+- @crossmint/common-sdk-auth@1.0.18
+
+## 1.2.7
+
+### Patch Changes
+
+- @crossmint/common-sdk-auth@1.0.17
+
 ## 1.2.6
 
 ### Patch Changes